Reorganize source code into subdirectories
