+++ scripts/precompile.sh
@@ -13,7 +13,7 @@ log_exit() {
 start_time=$(date +%s)
 
 PLUGIN_NAME=Phantom
-BUILD_TYPE=Release
+BUILD_TYPE=Debug
 FILENAME=${PLUGIN_NAME}Data.h
 
 for i in "$@"; do
@@ -42,7 +42,7 @@ build_plugin_binaries
 copy_header_file() {
     rm -f src/${FILENAME}
 
-    cp -f bin/juce_binarydata_${PLUGIN_NAME}Data/JuceLibraryCode/${FILENAME} src/${FILENAME} || log_exit "\n[Error] Failed to copy header file\n"
+    cp -f bin/juce_binarydata_${PLUGIN_NAME}Data/JuceLibraryCode/${FILENAME} src/utils/${FILENAME} || log_exit "\n[Error] Failed to copy header file\n"
     echo -e "[Success] Copied generated header file to src directory!\n"
 }
