+++ src/components/PhantomAnalyzer.h
