+++ src/effects/PhantomAmplifier.h
