+++ src/effects/PhantomFilter.h
@@ -13,6 +13,7 @@
 #pragma once
 
 #include "JuceHeader.h"
+
 #include "PhantomWaveshaper.h"
 
 /**
