+++ src/generators/PhantomEnvelopeGenerator.cpp
